orchestrator/pyproject.toml
[tool.poetry]
name = "mcp-drug-discovery-orchestrator"
version = "0.1.0"
description = "Drug discovery pipeline orchestrator using MCP servers"
authors = ["Your Name <your.email@example.com>"]
readme = "../README.md"
packages = [{include = "orchestrator", from = "."}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
typer = "^0.9.0"
httpx = {extras = ["http2"], version = "^0.25.2"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
pandas = "^2.1.4"
numpy = "^1.24.4"
pyyaml = "^6.0.1"
networkx = "^3.2.1"
rdkit-pypi = "^2023.9.2"
scikit-learn = "^1.3.2"
loguru = "^0.7.2"
tqdm = "^4.66.1"
redis = "^5.0.1"
aiofiles = "^23.2.0"
plotly = "^5.17.0"
biopython = "^1.82"
requests = "^2.31.0"
openpyxl = "^3.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pre-commit = "^3.6.0"

[tool.poetry.scripts]
mcp-orchestrator = "orchestrator.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=orchestrator --cov-report=html --cov-report=term-missing"
